requirements.txt
httpx>=0.23.0
mcp-server-sdk>=0.1.0
pytest>=7.0.0
pytest-asyncio>=0.20.0
python-dotenv>=1.0.0